Edit getSound

+++ includes/client/DecodedSound.hpp
@@ -1,10 +1,12 @@
 #ifndef BABEL_DECODEDSOUND_H
 #define BABEL_DECODEDSOUND_H
 
+#include <vector>
+
 namespace Babel {
 	typedef struct DecodedSound {
 		std::vector<float>	buffer;
-		int					size = 0;
-	}			   DecodedSound;
+		unsigned long		size = 0;
+	}			DecodedSound;
 } // namespace Babel
 #endif
